refactor(lib): import MODEL_FAMILIES from the model-map leaf below storage
Review follow-up: migrations.ts (and the same pattern in accounts.ts and
codex-cli/sync.ts) imported MODEL_FAMILIES/ModelFamily via the
prompts/codex.ts re-export, which sits above the storage layer in the
documented hierarchy. Retargeted to the model-map leaf the symbols were
moved to, so a future back-import into the storage layer cannot silently
re-open a cycle through these edges.
